requirements.txt
httpx>=0.27.0
beautifulsoup4>=4.12.0
pdfplumber>=0.11.0
pyyaml>=6.0.0
python-dotenv>=1.0.0
loguru>=0.7.0
jinja2>=3.1.0
